separate the sx styles from the (dashboard) layout file
