+++ src/app/[locale]/(dashboard)/layout.tsx
@@ -4,21 +4,7 @@ import Topbar from './components/Topbar/Topbar';
 import Sidebar from './components/Sidebar/Sidebar';
 import Main from './components/Main/Main';
 import SavedData from './components/savedData/SavedData';
-
-const paddings = { xs: 1, md: 2 };
-const sxStyles = {
-  app: {
-    height: '100%',
-    display: 'flex',
-    p: paddings,
-  },
-  board: {
-    overflow: 'auto',
-    flexGrow: 1,
-    display: 'flex',
-    flexDirection: 'column',
-  },
-};
+import sxStyles from './styles';
 
 export default function Dashboard(props: { children: React.ReactNode }) {
   return (
